+++ docs/source/conf.py
@@ -27,7 +27,7 @@ author = 'Dany Cajas'
 # The short X.Y version
 version = 'latest'
 # The full version, including alpha/beta/rc tags
-release = '3.2.0'
+release = '3.3.0'
 
 
 # -- General configuration ---------------------------------------------------