+++ riskfolio/__init__.py
@@ -16,6 +16,7 @@ from .Reports import *
 from .AuxFunctions import *
 from .DBHT import *
 from .OwaWeights import *
+from .GerberStatistic import *
 
 
-__version__ = "3.2.0"
+__version__ = "3.3.0"