+++ CHANGELOG.rst
@@ -7,6 +7,12 @@ Changelog
     <a href='https://ko-fi.com/B0B833SXD' target='_blank'><img height='36'style='border:0px;height:36px;' src='https://cdn.ko-fi.com/cdn/kofi1.png?v=2' border='0' alt='Buy Me a Coffee at ko-fi.com' /></a>
 
 
+Version 3.3.0
+=============
+
+- Adds Kendall Tau and Gerber statistic as options for codependence matrix in HCPortfolio object.
+- Adds Gerber statistic as an option for covariance matrix estimator in Portfolio and HCPortfolio objects.
+
 Version 3.2.0
 =============